ScoutBH/requirements.txt
discord.py==1.7.3
requests
asyncio
Pillow
pandas
seaborn
matplotlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pymongo[srv]==3.12.0